src/services/business/riskManager.ts
import { ethers } from 'ethers';
import { ValidationError } from '@/utils/errors';
import { EventLogger } from '../audit/eventLogger';

export class RiskManager {
  private static readonly HIGH_RISK_THRESHOLD = ethers.utils.parseUnits('50000', 6);
  private static readonly SUSPICIOUS_PATTERNS = [
    'multiple_failed_attempts',
    'rapid_withdrawals',
    'unusual_ip_changes'
  ];

  private eventLogger: EventLogger;

  constructor() {
    this.eventLogger = EventLogger.getInstance();
  }

  async assessTransactionRisk(
    userId: string,
    amount: ethers.BigNumber,
    transactionHistory: any[]
  ): Promise<{ riskLevel: 'low' | 'medium' | 'high'; requiresApproval: boolean }> {
    const riskFactors = await Promise.all([
      this.checkAmount(amount),
      this.checkTransactionPattern(transactionHistory),
      this.checkUserHistory(userId)
    ]);

    const riskLevel = this.calculateRiskLevel(riskFactors);
    const requiresApproval = riskLevel === 'high';

    if (requiresApproval) {
      await this.eventLogger.logSecurityEvent(
        userId,
        'HIGH_RISK_TRANSACTION',
        { amount: amount.toString(), riskFactors }
      );
    }

    return { riskLevel, requiresApproval };
  }

  private async checkAmount(amount: ethers.BigNumber): Promise<number> {
    return amount.gte(this.HIGH_RISK_THRESHOLD) ? 1 : 0;
  }

  private async checkTransactionPattern(history: any[]): Promise<number> {
    // Implement pattern recognition logic
    return 0;
  }

  private async checkUserHistory(userId: string): Promise<number> {
    // Implement user history check
    return 0;
  }

  private calculateRiskLevel(factors: number[]): 'low' | 'medium' | 'high' {
    const totalRisk = factors.reduce((sum, factor) => sum + factor, 0);
    if (totalRisk >= 2) return 'high';
    if (totalRisk === 1) return 'medium';
    return 'low';
  }
}